add component home e seus botoes de ir para compras e ir vender funcionarem
